Give each scheduled job back to the domain that owns it
`schedulerService.js` was 1,543 lines of unrelated work sitting directly
beside `services/scheduler/index.js`, whose own header says it is a
registry. Wars, economy seasons, ranked seasons, bank interest,
pet-of-the-week, shop price recalculation, the market expiry sweep and the
weekly champions were all in the one module, so a change to war logic was
made in the same file as bank interest and the whole thing was one review
surface. Ten of the registry's entries named it as their `service`, which
meant /health and every dead-letter entry pointed at a file that could not
say which subsystem had failed.

Each job's body now lives in the service that owns its domain — warService,
economySeasonService, leaderboardBadgeService, weeklyChampionService,
shopPricingService, rankedSeasonService, bankService, marketService, and
Pet of the Week into petService beside the pet mechanics it sets. The
registry points at those and holds no bodies of its own;
`postScheduledNewspapers` reaches newspaperService directly rather than
through a re-export. The two helpers every job shared move down a layer:
`postAnnouncement` into utils/guildAnnounce.js, and `owedSummary` into
utils/owedPayout.js beside `recordOwedPayout`, which its two callers
already use.

No behaviour changes and no scheduling changes. The `service` recorded on a
run and on an owed-payout record is now the domain service rather than
`schedulerService`, which is the point of the split — tests/schedulerBootstrap
holds `service` to the module each `fn` actually requires, and holds every
job body outside the registry directory.

Closes #931
