+++ tests/guildIndexes.test.js
@@ -118,16 +118,16 @@ describe('each scheduled Guild sweep has an index matching its filter', () => {
             partial: { 'tempVoice.enabled': true },
         },
         {
-            what: 'schedulerService.recalcShopPrices',
-            source: 'services/schedulerService.js',
+            what: 'shopPricingService.recalcShopPrices',
+            source: 'services/shopPricingService.js',
             filter: "Guild.find({ 'dynamicPricing.enabled': true }",
             index: 'idx_guilds_dynamic_pricing',
             keys: { 'dynamicPricing.enabled': 1 },
             partial: { 'dynamicPricing.enabled': true },
         },
         {
-            what: 'schedulerService bank-district payout',
-            source: 'services/schedulerService.js',
+            what: 'bankService bank-district payout',
+            source: 'services/bankService.js',
             filter: "$elemMatch: { districtId: 'bank', activeUntil: { $gt: now } }",
             index: 'idx_guilds_district_active',
             keys: { 'districts.districtId': 1, 'districts.activeUntil': 1 },
