+++ tests/schedulerBootstrap.test.js
@@ -21,6 +21,26 @@ describe('scheduler job table', () => {
         expect(new Set(names).size).toBe(names.length);
     });
 
+    // #931. `service` is what runJob records a run under, what /health reports
+    // and what a dead-letter entry is filed against, so a job whose service
+    // name does not match the module it actually calls sends whoever reads the
+    // failure to the wrong file. Ten of these named `schedulerService` and
+    // pointed at it regardless of whether the function inside was about wars,
+    // bank interest or shop prices; now each names the domain service that owns
+    // the job, and this is what stops the two drifting apart again.
+    test.each(JOBS.map(j => [j.name, j]))('%s: service names the module its fn requires', (_name, job) => {
+        const match = /require\('([^']+)'\)/.exec(job.fn.toString());
+        expect(match).not.toBeNull();
+        expect(job.service).toBe(match[1].split('/').pop());
+    });
+
+    // The registry may point at a domain service; it may not become one. A job
+    // whose body sits inside services/scheduler/ is the grab-bag returning
+    // under a new name — every `fn` reaches out of this directory, not into it.
+    test.each(JOBS.map(j => [j.name, j]))('%s lives outside the registry', (_name, job) => {
+        expect(job.fn.toString()).toMatch(/require\('\.\.\/[^/']+'\)/);
+    });
+
     test('starter names are unique', () => {
         const names = STARTERS.map(s => s.name);
         expect(new Set(names).size).toBe(names.length);
